Rename purchasing api file
